Include "quad" fix; added UseBikeLogicOnQuads .ini option.
