+++ IniReader/LICENSE.txt
