--- RealTrafficFix/RealTrafficFix.vcxproj.filters
@@ -1,6 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Project ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <ItemGroup>
-    <ClCompile Include="RealTrafficFix.cpp" />
-  </ItemGroup>
-</Project>
